% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciber.R
\name{ciber}
\alias{ciber}
\title{CIBERSORT评估高低风险分组免疫细胞浸润差异}
\usage{
ciber(exp, dat, fileName1, fileName2)
}
\arguments{
\item{exp}{表达谱}

\item{dat}{sample status os Risk group}

\item{fileName1}{文件名}

\item{fileName2}{图片名}
}
\value{
文本和图片
}
\description{
CIBERSORT评估高低风险分组免疫细胞浸润差异
}
\examples{
ciber(exp,dat,fileName1,fileName2)
}
