% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_model.R
\name{clinBox}
\alias{clinBox}
\title{风险得分在临床特征间的分布箱线图}
\usage{
clinBox(feature, clintype, fileName)
}
\arguments{
\item{feature}{临床特征，Risk group}

\item{clintype}{临床特征名字}

\item{fileName}{图名}
}
\value{
图
}
\description{
风险得分在临床特征间的分布箱线图
}
\examples{
clinBox(feature,clintype,fileName)
}
