% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{consensus}
\alias{consensus}
\title{利用指定基因进行一致性聚类，返回分类结果}
\usage{
consensus(
  exp,
  genelist,
  pItem = 0.8,
  pFeature = 0.8,
  clusterAlg = "pam",
  distance = "spearman"
)
}
\arguments{
\item{exp}{:表达谱数据}

\item{genelist}{：基因列表，如m6A}

\item{pItem}{:0.8}

\item{pFeature}{:0.8}

\item{clusterAlg}{：pam,hc}

\item{distance}{:euclidean,spearman,pearson}
}
\value{
：cluster1<-conClust[[k]]$consensusClass
}
\description{
利用指定基因进行一致性聚类，返回分类结果
}
\examples{
cluster<-consensus(exp,genelist,pItem = 0.8,pFeature = 0.8,clusterAlg = "pam",distance = 'spearman')
}
