% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyGO.R
\name{get_all_children}
\alias{get_all_children}
\title{get_all_children}
\usage{
get_all_children(parent_go, ko_go = character(), depth = 1, as.list = FALSE)
}
\arguments{
\item{parent_go}{GO to find children for.}

\item{ko_go}{These GO terms can't be returned as children.  Mainly used recursively by function and should not be supplied by user.}

\item{depth}{Current depth of recursion.  Mainly used recursively by function and should not be supplied by user.}

\item{as.list}{If TRUE, a list with children per input parent_go is returned instead of unique character vector for all parent_go.}
}
\value{
Either a character vector of children (includes parents). If as.list == TRUE, a list per parent_go.
}
\description{
get_all_children
}
\examples{
go_ids = head(search_GO_by_term("MAPK"))
parent_go = get_parents(go_ids)
get_all_children(parent_go)
get_all_children(parent_go, as.list = TRUE)
}
