% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyGO.R
\name{get_gene_name2go}
\alias{get_gene_name2go}
\title{get_gene_name2go}
\usage{
get_gene_name2go(org.db = org.Hs.eg.db::org.Hs.eg.db)
}
\arguments{
\item{org.db}{OrgDb such as org.Hs.eg.db, org.Mm.eg.db, org.Dm.eg.db etc.}
}
\value{
data.frame mapping GO to SYMBOL
}
\description{
This function is used internally by other functions to create a data.frame mapping GO terms to gene_names.  May be called to pre-generate this data.frame and supplied to later functions to speed things up greatly.
}
\examples{
gene_name2go = get_gene_name2go(org.Hs.eg.db::org.Hs.eg.db)
}
