% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyGO.R
\name{get_GO_children}
\alias{get_GO_children}
\title{get_GO_children}
\usage{
get_GO_children(parent_go, as.list = FALSE)
}
\arguments{
\item{parent_go}{GO to find children for.}

\item{as.list}{If TRUE, a list with children per input parent_go is returned
instead of unique character vector for all parent_go.}
}
\value{
Either a character vector of children (includes parents). If as.list
  == TRUE, a list per parent_go.
}
\description{
get_GO_children
}
\examples{
go_ids = head(search_GO_by_term("MAPK"))
parent_go = get_GO_parents(go_ids)
get_GO_children(parent_go)
get_GO_children(parent_go, as.list = TRUE)
}
