% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyGO.R
\name{get_GO_gene_names}
\alias{get_GO_gene_names}
\title{get_GO_gene_names}
\usage{
get_GO_gene_names(
  go_ids,
  org.db = org.Hs.eg.db::org.Hs.eg.db,
  as.list = FALSE,
  gene_name2go = NULL
)
}
\arguments{
\item{go_ids}{character vector of GO ids, like GO:1234.}

\item{org.db}{OrgDb such as org.Hs.eg.db, org.Mm.eg.db, org.Dm.eg.db etc.}

\item{as.list}{If TRUE, a list with children per input parent_go is returned
instead of unique character vector for all parent_go.}

\item{gene_name2go}{Optional result of get_GO_gene_name2go() for org.db. Much
faster if supplied.}
}
\value{
data.frame with mapping for ENTREZID, GO, EVIDENCE, ONTOLOGY, SYMBOL
}
\description{
get_GO_gene_names
}
\examples{
gene_name2go = get_GO_gene_name2go(org.Hs.eg.db::org.Hs.eg.db)
head(gene_name2go)
}
