% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyGO.R
\name{get_GO_info}
\alias{get_GO_info}
\title{get_GO_info}
\usage{
get_GO_info(
  go_ids,
  include_children = FALSE,
  include_parents = FALSE,
  include_genes = FALSE,
  org.db = NULL,
  gene_name2go = NULL,
  sep = ","
)
}
\arguments{
\item{go_ids}{character vector of GO ids, like GO:1234.}

\item{include_children}{If TRUE, a column of CHILDREN is added. Values are
comma delimitted.}

\item{include_parents}{If TRUE, a column of PARENTS is added. Values are
comma delimitted.}

\item{include_genes}{If TRUE, a column of GENES is added. Values are comma
delimitted. OrgDb must be supplied.}

\item{org.db}{OrgDb such as org.Hs.eg.db, org.Mm.eg.db, org.Dm.eg.db etc.
Required for include_genes.}

\item{gene_name2go}{Optional result of get_GO_gene_name2go() for org.db. Much
faster if supplied.}

\item{sep}{character delimiter to use for items in extra columns. Default is ",".}
}
\value{
A data.frame of information about supplied GO ids.
}
\description{
get_GO_info
}
\examples{
go_ids = head(search_GO_by_term("MAPK"))
get_GO_info(go_ids)
# Additional columns can be added though these take a little longer.
get_GO_info(go_ids,
  include_children = TRUE,
  include_parents = TRUE,
  include_genes = TRUE,
  org.db = org.Hs.eg.db::org.Hs.eg.db)
# Speed things up by pre-generating gene_name2go and reusing.
gene_name2go = get_GO_gene_name2go(org.Hs.eg.db::org.Hs.eg.db)
get_GO_info(go_ids,
  include_children = TRUE,
  include_parents = TRUE,
  include_genes = TRUE,
  org.db = org.Hs.eg.db::org.Hs.eg.db,
  gene_name2go = gene_name2go)
}
