% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyGO.R
\name{get_GO_parents}
\alias{get_GO_parents}
\title{get_GO_parents}
\usage{
get_GO_parents(children_go, as.list = FALSE)
}
\arguments{
\item{children_go}{GO to find parents for.}

\item{as.list}{If TRUE, a list with children per input parent_go is returned
instead of unique character vector for all parent_go.}
}
\value{
GO ids of parent terms.
}
\description{
get_GO_parents
}
\examples{
go_ids = head(search_GO_by_term("MAPK"))
get_GO_parents(go_ids)
get_GO_parents(go_ids, as.list = TRUE)
}
