% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyGO.R
\name{get_parents}
\alias{get_parents}
\title{get_parents}
\usage{
get_parents(children_go)
}
\arguments{
\item{children_go}{GO to find parents for.}
}
\value{
GO ids of parent terms.
}
\description{
get_parents
}
\examples{
go_ids = head(search_GO_by_term("MAPK"))
get_parents(go_ids)
}
