% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClin.R
\name{getClin}
\alias{getClin}
\title{下载样本的临床信息}
\usage{
getClin(cancer, fileName)
}
\arguments{
\item{cancer}{："hnsc_tcga"}

\item{fileName}{：输出文件名称}
}
\value{
：result<-list(clin_down=clin_down,survival=tcga_pdata_os)
}
\description{
下载样本的临床信息
}
\examples{
getClin(cancer="hnsc_tcga",fileName="./result/Table/SupplementaryTable1_clin.txt")
}
