% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyFun.R
\name{getCors}
\alias{getCors}
\title{计算所有基因跟m6A基因的相关性}
\usage{
getCors(cancerExp, genelist)
}
\arguments{
\item{cancerExp}{:表达谱数据}

\item{genelist}{：基因列表}
}
\value{
:list(cors=cors,pvalues=pvalues)
}
\description{
计算所有基因跟m6A基因的相关性
}
\examples{
getCors(cancerExp=exp,genelist=m6A)
}
