% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyFun.R
\name{getCors_sig}
\alias{getCors_sig}
\title{筛选与m6A显著相关的基因}
\usage{
getCors_sig(result, cor, cor_percent, pvalue, pvalue_percent)
}
\arguments{
\item{result}{getCors返回的结果}

\item{cor}{相关性阈值}

\item{cor_percent}{比例}

\item{pvalue}{显著性阈值}

\item{pvalue_percent}{比例}
}
\value{
result<-list(cors=cors,pvalues=pvalues,genelist=cd_m6A_geneset1)
}
\description{
筛选与m6A显著相关的基因
}
\examples{
getCors_sig(result,cor=0.3,cor_percent=0.5,pvalue=0.05,pvalue_percent=1)
}
