% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRNA_series1.R
\name{getDEGs}
\alias{getDEGs}
\title{根据表达谱和样本分类信息进行差异基因分析}
\usage{
getDEGs(exp, pdata, caseLabel, controlLabel, fc, pvalue = NULL, adjP, fileName)
}
\arguments{
\item{exp}{：表达谱文件}

\item{pdata}{：样本分类信息}

\item{caseLabel}{：分组编号如Tumor}

\item{controlLabel}{：分组编号如Normal}

\item{fc}{:差异倍数，如2或者1.5}

\item{pvalue}{:是否使用pvalue,如果不使用，则pvalue=NULL}

\item{adjP}{:阈值，一般为0.05}

\item{fileName}{:nrDEG输出文件名}
}
\value{
:差异基因列表
}
\description{
根据表达谱和样本分类信息进行差异基因分析
}
\examples{
getDEGs(exp=result[[1]],pdata=result[[2]],caseLabel="Tumor",controlLabel="Normal"
,fc=1.5,pvalue=NULL,adjP=0.05,fileName="SupplementaryTable_diffgene.txt")

}
