% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRNA_series1.R
\name{getExp}
\alias{getExp}
\title{处理下载的表达谱数据}
\usage{
getExp(inputfile, gene_map_info_v22)
}
\arguments{
\item{inputfile}{xena下载的表达谱数据}

\item{gene_map_info_v22}{/Users/denglili/Desktop/database/gencode.v22.annotation.gene.probeMap}
}
\value{
list result<-list(exp=mrna_datexpr,pdata=pdata)
}
\description{
处理下载的表达谱数据
}
\examples{
getexp(inputfile="TCGA-LIHC.htseq_fpkm.tsv",gene_map_info_v22=
"/Users/denglili/Desktop/database/gencode.v22.annotation.gene.probeMap")
}
