% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{getKM}
\alias{getKM}
\title{绘制分组样本的KM曲线}
\usage{
getKM(cluster, clinFile, fileName, colors)
}
\arguments{
\item{cluster}{：分组信息，第一列样本，第二列分组}

\item{clinFile}{：临床信息，第一列样本，后面列包含time_os（months），status_os}

\item{fileName}{:KM曲线图名称}

\item{colors}{:线条配色}
}
\value{
：3张KM曲线图
}
\description{
绘制分组样本的KM曲线
}
\examples{
：getKM(cluster,clinFile,fileName="Figure1",colors=colorRampPalette(brewer.pal(5,"Set3")))
}
