% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_model.R
\name{getKMgene}
\alias{getKMgene}
\title{根据单个基因表达谱绘制KM曲线}
\usage{
getKMgene(exp, plotgene, clinFile, fileName, color = c("#E7B800", "#2E9FDF"))
}
\arguments{
\item{exp}{表达谱}

\item{plotgene}{要画图的基因}

\item{clinFile}{临床信息文件，包含sample time_os status_os}

\item{fileName}{图片文件名}

\item{color}{图片中曲线颜色，有默认值}
}
\value{
p.val
}
\description{
根据单个基因表达谱绘制KM曲线
}
\examples{
getKMgene(exp,plotgene,clinFile,fileName,color=c("#E7B800","#2E9FDF"))
}
