% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_model.R
\name{getKMriskscore}
\alias{getKMriskscore}
\title{根据风险得分进行分类，然后绘制KM曲线}
\usage{
getKMriskscore(
  merged_data,
  riskScore,
  fileName1,
  fileName2,
  color = c("firebrick3", "skyblue")
)
}
\arguments{
\item{merged_data}{包含time_os，status_os}

\item{riskScore}{第一列样本，第二列风险得分}

\item{fileName1}{文件名}

\item{fileName2}{图片名}

\item{color}{KM曲线颜色}
}
\value{
图和文件,以及dat
}
\description{
根据风险得分进行分类，然后绘制KM曲线
}
\examples{
getKMriskscore(merged_data,riskScore,fileName1,fileName2,color=c("firebrick3","skyblue"))
dat<-getKMriskscore(merged_data=merged_data,riskScore=riskScore,fileName1="./result/Table/SupplementaryTable14",fileName2="./result/Figure/Figure11D",color=c("firebrick3","skyblue"))
plotThree(dat=dat,sigData=sigData,fileName="./result/Figure/Figure11")
roc1(dat,marker,fileName)
roc2(dat,marker,fileName)
}
