% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_model.R
\name{getKMsubclin}
\alias{getKMsubclin}
\title{KM曲线临床稳定性分析}
\usage{
getKMsubclin(dat, subsamples, label, fileName)
}
\arguments{
\item{dat}{包括sample time status Risk group}

\item{subsamples}{临床亚群样本名称列表}

\item{label}{临床特征名称"age: >60" "Stage: I-II"}

\item{fileName}{图名}
}
\value{
KM曲线图
}
\description{
KM曲线临床稳定性分析
}
\examples{
getKMsubclin(dat,subsamples,label,fileName)
}
