% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna.R
\name{getModuleHub}
\alias{getModuleHub}
\title{筛选模块hub基因}
\usage{
getModuleHub(cut1, cut2, module = "turquoise", lastinput, fileName1)
}
\arguments{
\item{cut1}{横坐标阈值}

\item{cut2}{纵坐标阈值}

\item{module}{模型名称}

\item{lastinput}{myWGCNA输出结果}

\item{fileName1}{输出图片名称}
}
\value{
模块hub基因
}
\description{
筛选模块hub基因
}
\examples{
getModuleHub(cut1,cut2,module = "turquoise",lastinput,fileName1)
}
