% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna.R
\name{GS}
\alias{GS}
\title{计算基因网显著性}
\usage{
GS(trait = traitDt[, 3], lastinput, h, fileName1, fileName2)
}
\arguments{
\item{trait}{临床特征，选择其中一列}

\item{lastinput}{myWGCNA输出结果}

\item{h}{虚线高度}

\item{fileName1}{文本名字}

\item{fileName2}{图片名字}
}
\value{
图和文本
}
\description{
计算基因网显著性
}
\examples{
GS(trait=traitDt[,3],lastinput,h=0.8,fileName1,fileName1)
}
