% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyFun.R
\name{hrd}
\alias{hrd}
\title{risk score与突变负荷、HRD、新抗原负荷和染色体不稳定、干性指数的相关性分析}
\usage{
hrd(cancertype, fileName1, fileName2)
}
\arguments{
\item{cancertype}{癌症类型STAD}

\item{fileName1}{图片名称}

\item{fileName2}{文本名称}
}
\value{
图片和文本
}
\description{
risk score与突变负荷、HRD、新抗原负荷和染色体不稳定、干性指数的相关性分析
}
\examples{
hrd(cancertype,fileName1,fileName2)
}
