% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huoshantu.R
\name{huoshantu}
\alias{huoshantu}
\title{Title}
\usage{
huoshantu(nrDEG, fileName, width, height, label = TRUE)
}
\arguments{
\item{nrDEG}{: limma差异基因分析结果，增加change一列}

\item{fileName}{：输出文件名字}

\item{width}{：宽度}

\item{height}{：长度}

\item{label}{:火山图上是否标记top10的基因}
}
\value{
返回一张带标签的火山图
}
\description{
Title
}
