% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_model.R
\name{lasso}
\alias{lasso}
\title{lasso分析}
\usage{
lasso(merged_data, cox_gene, fileName, fileName1)
}
\arguments{
\item{cox_gene}{基因列表}

\item{fileName}{图片名称}

\item{fileName1}{文件名称}
}
\value{
coef_gene
}
\description{
lasso分析
}
\examples{
lasso(merged_data,cox_gene,fileName,fileName1)
}
