% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyFun.R
\name{myestimate}
\alias{myestimate}
\title{高低风险分组免疫评分、基质评分、肿瘤纯度差异}
\usage{
myestimate(exp, dat, fileName1, fileName2)
}
\arguments{
\item{exp}{表达谱}

\item{dat}{sample statu os Risk group}

\item{fileName1}{文本文件}

\item{fileName2}{图片名称}
}
\value{
文本和图片
}
\description{
高低风险分组免疫评分、基质评分、肿瘤纯度差异
}
\examples{
myestimate(exp,dat,fileName1,fileName2)
}
