% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRNA_series1.R
\name{myvenn}
\alias{myvenn}
\title{绘制韦恩图1}
\usage{
myvenn(x, fileName)
}
\arguments{
\item{x}{list列表}

\item{fileName}{图片名字}
}
\value{
韦恩图
}
\description{
绘制韦恩图1
}
\examples{
myvenn(x=list1,fileName="./result/Figure/Figure-300ppi.tif")
}
