% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna.R
\name{myWGCNA}
\alias{myWGCNA}
\title{WGCNA模块挖掘}
\usage{
myWGCNA(exp, genelist, fileName, fileName1)
}
\arguments{
\item{exp}{基因表达谱}

\item{genelist}{基因列表}

\item{fileName}{图片文件名称前缀}

\item{fileName1}{文本文件名称前缀}
}
\value{
result<-list(datexpr=datexpr,moduleColors=moduleColors,pow=pow)
}
\description{
WGCNA模块挖掘
}
\examples{
myWGCNA(exp,genelist,fileName="./result/Figure/Figure5",fileName1="./result/Table/SupplementaryTable")
mrna_datexpr_cancer<-Data1[[1]][,Data1[[2]][which(Data1[[2]][,2]=="Tumor"),1]]
mylist<-unique(c(m6A,cd_m6A_geneset1,cd_m6A_geneset2))
aa<-myWGCNA(exp=mrna_datexpr_cancer,genelist=mylist,fileName="./result/Figure/Figure5",fileName1="./result/Table/SupplementaryTable5")
}
