% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrich.R
\name{pathway_enrich}
\alias{pathway_enrich}
\title{对基因进行功能富集分析}
\usage{
pathway_enrich(genelist, fileName, width, height)
}
\arguments{
\item{genelist}{输入基因列表}

\item{fileName}{输出结果文件}

\item{width}{保存图片的宽度}

\item{height}{保存图片的高度}
}
\value{
气泡图和富集结果csv文件
}
\description{
对基因进行功能富集分析
}
\examples{
pathway_enrich(genelit=m6A,fileName="aa",width=10,height=10)
}
