% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRNA_series1.R
\name{pca1}
\alias{pca1}
\title{PCA分析，不带圈圈}
\usage{
pca1(exp, pdata, fileName, color = c("#DFC03F", "#4376BE"))
}
\arguments{
\item{exp}{表达谱数据}

\item{pdata}{分组信息}

\item{fileName}{图片文件名称}

\item{color}{分组颜色，NULL}
}
\value{
PCA图
}
\description{
PCA分析，不带圈圈
}
\examples{
pca1(exp,pdata,fileName,color=c("#DFC03F","#4376BE"))
}
