% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRNA_series1.R
\name{pca2}
\alias{pca2}
\title{PCA分析，带圈圈}
\usage{
pca2(exp, pdata, fileName)
}
\arguments{
\item{exp}{表达谱数据}

\item{pdata}{分组信息}

\item{fileName}{图片文件名称}
}
\value{
PAC图
}
\description{
PCA分析，带圈圈
}
\examples{
pca2(exp,pdata,fileName)
}
