% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRNA_series1.R
\name{phmap}
\alias{phmap}
\title{绘制热图}
\usage{
phmap(
  gg,
  anno_col = NULL,
  ann_colors = NULL,
  bk = seq(-2.5, 2.5, length = 100),
  color = colorRampPalette(c("navy", "white", "firebrick3"))(100),
  fileName,
  showrownames = T,
  showcolnames = T,
  width,
  height
)
}
\arguments{
\item{gg}{绘图数据}

\item{anno_col}{colbar列注释}

\item{ann_colors}{列注释的颜色}

\item{bk}{分段}

\item{color}{热图颜色}

\item{fileName}{热图保存文件名}

\item{showrownames}{是否显示行名T}

\item{showcolnames}{是否显示列名T}

\item{height}{}
}
\value{
热图
}
\description{
绘制热图
}
\examples{
phmap(gg,anno_col=NULL,ann_colors=NULL,bk=seq(-2.5,2.5,length=100),color=colorRampPalette(c("navy", "white", "firebrick3"))(100),fileName="./resut/Figure/Figure")
}
