% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_model.R
\name{plotForest}
\alias{plotForest}
\title{对筛选的基因绘制森林图}
\usage{
plotForest(coxResult, xticks, fileName)
}
\arguments{
\item{coxResult}{singleCox分析结果}

\item{xticks}{x轴区间}

\item{fileName}{图片名称}
}
\value{
森林图
}
\description{
对筛选的基因绘制森林图
}
\examples{
plotForest(coxResult,xticks,fileName)
}
