% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_model.R
\name{plotThree}
\alias{plotThree}
\title{绘制三联图}
\usage{
plotThree(dat, sigData, fileName)
}
\arguments{
\item{dat}{包括sample time status Risk group}

\item{sigData}{表达谱}

\item{fileName}{图名}
}
\value{
三联图ABC
}
\description{
绘制三联图
}
\examples{
plotTree(dat,sigData,fileName)
}
