% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_model.R
\name{roc1}
\alias{roc1}
\title{绘制ROC曲线}
\usage{
roc1(dat, marker, fileName)
}
\arguments{
\item{dat}{包括sample time status Risk group}

\item{marker}{1或者0，1表示利用风险得分，0表示利用分组}

\item{fileName}{图名}
}
\value{
1，3，5年的roc曲线
}
\description{
绘制ROC曲线
}
\examples{
roc1(dat,marker,fileName)
}
