% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_model.R
\name{roc2}
\alias{roc2}
\title{绘制单线条的ROC曲线}
\usage{
roc2(dat, marker, fileName)
}
\arguments{
\item{dat}{包括sample time status Risk group}

\item{marker}{"high"或者"low"}

\item{fileName}{图名}
}
\value{
ROC曲线
}
\description{
绘制单线条的ROC曲线
}
\examples{
roc2(dat,marker,fileName)
}
