% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_model.R
\name{roc3}
\alias{roc3}
\title{针对单个基因绘制一年的ROC曲线}
\usage{
roc3(dat, marker, fileName)
}
\arguments{
\item{dat}{sample status os Risk group}

\item{marker}{1:风险得分 0:Low}

\item{fileName}{图片名字}
}
\value{
图片
}
\description{
针对单个基因绘制一年的ROC曲线
}
\examples{
roc3(dat,marker,fileName)
}
