% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna.R
\name{selectModule}
\alias{selectModule}
\title{筛选模块特征}
\usage{
selectModule(lastinput, traitDt, fileName1, fileName2)
}
\arguments{
\item{lastinput}{：myWGCNA返回的结果}

\item{traitDt}{：临床特征表格}

\item{fileName1}{：文件名称}

\item{fileName2}{：图片名称}
}
\value{
图片和文本
}
\description{
筛选模块特征
}
\examples{
selectModule(lastinput,traitDt,fileName1="./result/Table/SupplementaryTable",fileName2="./result/Figure/Figure5")
traitDt<-as.matrix(clin_infor[match(rownames(aa$datexpr),rownames(clin_infor)),c(21,22,52,53,82)])
traitDt[,ncol(traitDt)]<-plyr::revalue(traitDt[,ncol(traitDt)],c("m6A.clusterA"=1,"m6A.clusterB"=2))
selectModule(lastinput=aa,traitDt=traitDt,fileName1="./result/Table/SupplementaryTable5",fileName2="./result/Figure/Figure5")
}
