% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_model.R
\name{singleCox}
\alias{singleCox}
\title{对候选基因进行单因素cox分析}
\usage{
singleCox(exp, candidate, clinFile, pvalue, fileName)
}
\arguments{
\item{exp}{表达谱}

\item{candidate}{候选基因}

\item{clinFile}{临床信息文件，包含sample time_os status_os}

\item{pvalue}{0.05}

\item{fileName}{输出结果文件}
}
\value{
筛选的与预后显著相关的基因
}
\description{
对候选基因进行单因素cox分析
}
\examples{
singleCox(exp,candidate,clinFile,pvalue,fileName)
}
