% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRNA_series1.R
\name{huoshantu}
\alias{huoshantu}
\title{绘制火山图}
\usage{
huoshantu(nrDEG, fileName, width, height, label = TRUE)
}
\arguments{
\item{nrDEG}{：getDEGs生成的差异基因分析文件}

\item{fileName}{：火山图文件名}

\item{width}{：宽度}

\item{height}{：长度}

\item{label}{：是否增加top10基因的标签}
}
\value{

}
\description{
绘制火山图
}
\examples{
huoshantu(nrDEG="aa.txt",fileName="Figure12A",width=8,height=8,label=TRUE)
}
